Update: Delegated parts of the authorization flow to JS
Callback is now handled in the frontend. The frontend then call the
exchange route with the authorization code as query param. The
authorization code is exchanged for access token and the required
information is acquired from the appropriate APIs. This is then sent
back to the frontend. The frontend never gets the access token, only the
information it needs.